requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
PyJWT==2.6.0
cryptography==40.0.0
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==11.2.1
django-filter==23.1
drf-spectacular==0.26.1
celery==5.2.7
redis==4.5.4
requests==2.28.2
faker==18.4.0
python-dotenv==1.1.0
dj-database-url==3.0.0
django-oauth-toolkit==3.0.1
whitenoise==6.9.0